Refactoring of NumberTag.
Added requires clause to problematic Constructor overload. The name constructor will be disabled if the template type T matches std::string. This allows us to use the same template class for String Tags. Therefore the NumberTag was renamed to ValueTag since it is not only used for numbers. Existing tags were updated accordingly.